refactor: Chain 기준 RPC API와 단독 RPC API 분리
refactor: RPC 관련 API들이 ID를 반환하도록 수정
