+++ src/main/java/com/baekho/bridgenet/domain/chain/dto/request/RpcAddRequestDTO.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 
 @Getter
 public class RpcAddRequestDTO {
+    private Long id;
     private String serviceName;
     private String url;
     private Protocol protocol;
