+++ src/main/java/com/baekho/bridgenet/domain/chain/dto/response/RpcResponseDTO.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class RpcResponseDTO {
+    private Long id;
     private String serviceName;
     private String url;
     private Protocol protocol;
