+++ src/main/java/com/baekho/bridgenet/domain/chain/dto/response/RpcUpdateResponseDTO.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class RpcUpdateResponseDTO {
+    private Long id;
     private String serviceName;
     private String url;
     private Protocol protocol;
