+++ src/main/java/com/baekho/bridgenet/domain/chain/service/ChainService.java
@@ -29,7 +29,6 @@ import org.web3j.protocol.Web3j;
 import org.web3j.protocol.core.DefaultBlockParameterName;
 import org.web3j.protocol.core.methods.response.EthGetBalance;
 import org.web3j.protocol.core.methods.response.TransactionReceipt;
-import org.web3j.protocol.http.HttpService;
 import org.web3j.utils.Convert;
 
 import java.io.IOException;
